separated tracking to another file
