+++ src/main.js
@@ -27,41 +27,7 @@ import 'bootstrap/dist/css/bootstrap.css';
 import './main.scss';
 
 // google analytics
-import GoogleAnalytics from 'react-ga';
-
-GoogleAnalytics.initialize('UA-118469639-1');
-
-const withTracker = (WrappedComponent, options = {}) => {
-  const trackPage = page => {
-    GoogleAnalytics.set({
-      page,
-      ...options,
-    });
-    GoogleAnalytics.pageview(page);
-  };
-
-  const HOC = class extends React.Component {
-    componentDidMount() {
-      const page = this.props.location.pathname;
-      trackPage(page);
-    }
-
-    componentWillReceiveProps(nextProps) {
-      const currentPage = this.props.location.pathname;
-      const nextPage = nextProps.location.pathname;
-
-      if (currentPage !== nextPage) {
-        trackPage(nextPage);
-      }
-    }
-
-    render() {
-      return <WrappedComponent {...this.props} />;
-    }
-  };
-
-  return HOC;
-};
+import withTracker from './container/analytics.js';
 
 class PageTracker extends React.Component {
     render() { return null; }
